Check that opened indices have the right options
Turns out that libspatialindex will just ignore parameters that were
specified differently than what was specified previous. Ie, if you
create an index and then attempt to open it with different parameters
it'll just re-use the old parameters without causing an error. This
checks as many of the parameters as possible when reopening.