Replace TokyoCabinet with LevelDB
This is ground work for storing spatial data in the same index as the
rest of the key/value data. This is important so that index updates are
atomic. Having two separate storage systems with no interaction leaves
us open to the possibility of having indices that are inconsistent
(between the two, not necessarily in either one individually).

LevelDB also has the happy benefits of being both currently maintained
and non-GPL licensed. There's also the fact that it does background
compactions as well which saves us from having to worry about that.