Implement transactions to ensure atomic updates
This implements a basic transactional API ontop of the io::Storage class
using the leveldb::WriteBatch as a backing. Its a bit confusing that
writes are proxied through the Storage class back to the Transaction
class but this is due to how the storage manager works. It also ends up
preserving a super clean API for client code.