Implemented the query command
This accepts a WKB and a few other parameters and returns a list of hits
that meet the requested criteria. There last remaining major issue with
this code is the necessisty to add a custom index vistor to
easton::Index so that we're not loading all false negatives into RAM for
each query.

We'll also need to accept the other possible shape inputs instead of
just the WKB as well as implement the CRS translations once the rest of
that code exists.