Fix database reads during transactions
This bug occurred when some code attempted to read back key/value pairs
that it had written during the transaction. Since they weren't part of
the database it would come back empty.

This change just adds a hash map to the transaction that is checked for
the presence of a key. If the key doesn't exist in the transaction then
the request is forwarded to the backing database.