Set the write through flag on the index
It turns out that libspatialindex will do a lot of buffering during
indexing. Its then possible that some of these dirty pages are evicted
during a search when there's no write transaction in progress which
causes an error.

Setting write through ensures that all data is written within the
context of a transaction.