Implement the various shape queries
This extends the geometry functions to be able to accept a number of
shape definitions and then creates geometries based on those
definitions. Some of the shape definitions use some of the fancier geo
algorithms when creating shapes. For instance circles and ellipses
include calculations for adjusting to the ellipsoids they're operating
on.

All of these calculations were based directly on the existing code in
erl_spatial.