Limit exposure to spatialindex internals
This moves functions that are touching spatialindex internals to their
own source file. The need to access these mostly arrises from the need
to get the PropertySet from the Index object that wraps the actual RTree
index objects. For some reason it makes a copy during construction that
is not accessible through the public C API.