Fix index persistence
In order to be able to reopen an index we need to have access to its
index id. This is set after index creation so we have to reach into the
correct PropertySet to grab the value. When we reopen the index we just
need to specify the index id to use.

Its important to note that we can't just assign and index id as a
constant as spatialindex's logic assumes that if an id is passed in that
the index exists.