Implement result set bookmarks
This allows clients to access pages of results from the index based on a
limit and bookmark value. The bookmark value is a 2-tuple that contains
a document id and distance returned from a query with the same specified
shape. This work removes support for offsets which really didn't make
much sense without a defined ordering to the results anyway.