Implement disk and data sizes
These roughly parallel our notion of disk and data sizes for other
indices. The major difference is that the data size calculation is after
compression of user data. It wouldn't be impossible to track
uncompressed sizes in the future with some extra bookkeeping but for now
the rough approximation exposed by LevelDB should be close enough for
all intents and purposes.