Set a flush callback for the storage manager
Managed to waste upwards of ten hours debugging a segfault due to C++
memory initialization (or lack thereof). Triggering this happened to
coincide with adding some new CsMap usage so I was focused on that bit
of code as well as attemtping to use the many malloc debugging tools.
All better now though.