Avoid using CSepsg2adskCS when possible
The first time this mapping function is called takes roughly 1.5s on
this newer MacBook Pro. Rather than suffer that penalty each time an
index needs to be opened I've created a basic direct lookup table for
some of the more common EPSG codes. If a code doesn't exist in the table
we just fall back to the original CSepsg2adskCS function so that our
behavior is identical in the worst case.