Fix easton_geojson:to_wkb/1 performance
The process of prepending binaries is apparently fairly terrible
performance wise which is fairly obvious in hindsight. A recent customer
was loading some super large GeoJSON documents which ended up causing
the old binary concatenation version to allocate many gigabytes of
binaries for the conversion.

This just changes the accumulation logic to use an io_list() which is
then converted to a single binary at the very in with a single call to
iolist_to_binary/1.

BugzId: 42924