Open the libspatialindex before each request
This adds a reset function so that we recreate the index structure
before each request to the index. The maintainer of libspatialindex says
that they generally use these as short term objects so I'm guessing that
we might need to try and minimize the reuse of various buffers.

This change is motivated by a number of double delete errors that end up
resetting indexes completely and partially depending on when they happen
in an update. Under minimal load they appear to be happening around
twenty times a day.

If libspatialindex works out the bugs in their memory handling we'll
want to revert this so that we make better use of the various caches and
buffering that it provides.

BugzId: 43481