Only validate the index when opened the first time
This adds the abilityt to reset the index without revalidating it.
Validation requires reading the entire tree from disk and examining each
node in the tree. Obviously not something that is performant.

It does still maintain the initial validation when opening an index so
we keep the same level of safety as before we started resetting the
index object.

BugzId: 43841