Optimize index traversal
This patch changes how we traverse RTree's for intersection (bbox style)
queries. The standard method for traversal does not account for the fact
that we're paging through the result set. This method uses that
information to be able to skip traversing entire subtrees that only
contain points that are guaranteed to be further away than the last
entry on the page.

A further optimization would be to perform a similar optimization to
skip subtrees that only contain results that occur before the current
page based on the provided bookmark.

BugzId: 45296