Fix the ordering of Hit objects
There were a few subtle bugs when comparing Hit instances. First, the
HitCmp comparison should've been a strictly less than comparison but it
was implemnting less than or equal purely on distance. There was also a
bug when comparing docids when they were both empty as well as when one
was an exact prefix of the other.

This change moves the HitCmp to an overloaded less than operator for the
Hit class and fixes the bugs mentioned above. Secondly, it implements
the less than or equal operator as well for filtering values for
bookmarks.

BugzId: 52940