Corrected radius search error when used on dbs in equiangular projs.
The error is due to the implementation of radius searching. To perform a
radius search a geometry is created of a circle with a defined radius in
degrees based on the search radius expressed in metres. To calculate this
distance in degrees a calculation based on a point due east of the
central point of the search is made. This is where the error lies. At
high latitudes this distance in degrees is much greater in longitude than
it is in latitude, when the resultant radius is used the geometry is
therefore highly elliptical, with its semi-major axis aligned N-S,
therefore sampling more than the radius specified in that dimension than
the W-E dimension.

The fix is to to use the make_ellipse method in radius searching, this
calculates and uses a radius in both longitude and latitude which produces
correct results in equiangular projections.

BugzID: 56839