Fix bug in io.cc read_bytes
The bug was caused by passing a pointer from a different sized integer. When ei_decode_binary overwrote the value it was inadvertently overwriting other values in memory which on one platform ended up causing a smart pointer value to be set to null. This is a bug on all platforms where a long is bigger than an int32_t.

We can just set that parameter to NULL because on inspection of ei_decode_binary and ei_get_type. it's being set to the same value each time. Using a long and capturing the correct value would only allow for an assertion which we don't perform so there's no reason to capture the value.

NB: this is actually Paul Davis' fix that I am transcribing into this PR

BugzID: 61810